EWPP-220: Code structure fixes.
